refactor: merge dorm selection into dorm task
